Update to work with Channels version 2
There are two main changes:

 1. Use the new interface for sending layer messages.

 2. Remove lost task detection.

The second change is required due to there being multiple tasks able to be
run per worker, which breaks publishing the current set of tasks. For the time
being I'm going to leave the system as is and write some management commands to
help us locate and correct lost tasks. I have an alternative in mind for
automatically managing this, but it'll take a bit of work.